Move theme to layout-minor fixes
