--- pages/projects/ProjectList.module.css
@@ -1,4 +1,0 @@
-.container {
-  display: flex;
-  flex-wrap: wrap;
-}
